+++ tests/tester.js
@@ -176,6 +176,15 @@ const TESTER = (() => { //constructor factory
 				compare: 'pancake/',
 				empty: false,
 			},
+			{
+				name: 'Test14: Compile test with three modules from two organisations',
+				folder: 'test14',
+				script: {
+					options: [ '--org', '@nsw.gov.au @gov.au' ],
+				},
+				compare: 'pancake/',
+				empty: false,
+			},
 		],