+++ packages/pancake-react/CHANGELOG.md
@@ -6,6 +6,7 @@ Pancake React plugin
 
 ## Versions
 
+* [v1.2.0 - Multiple organisations](v120)
 * [v1.1.5 - Update dependencies](v115)
 * [v1.1.4 - Update dependencies](v114)
 * [v1.1.3 - Update node modules](v113)
@@ -17,6 +18,11 @@ Pancake React plugin
 
 
 ----------------------------------------------------------------------------------------------------------------------------------------------------------------
+
+## v1.2.0
+
+- Multiple organisations
+
 
 ## v1.1.5